Rename KernelSelection to JupyterInfo
